+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/Payload.java
@@ -45,6 +45,7 @@ public class Payload {
     public static final HeaderKey<ProcessKind> PROCESS_KIND = HeaderKey.register("_processKind", ProcessKind.class);
     public static final HeaderKey<Set<String>> PROCESS_TAGS = HeaderKey.registerSet("_processTags");
     public static final HeaderKey<Set<String>> OUT_EXPRESSIONS = HeaderKey.registerSet("_outExpr");
+    public static final HeaderKey<Map<String, Object>> REQUIREMENTS = HeaderKey.registerMap("_requirements");
 
     public static final AttachmentKey WORKSPACE_ARCHIVE = AttachmentKey.register("archive");