+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- concord-server, concord-agent: initial support for environmental
+`requirements` and agent capabilities.
+
 ### Changed
 
 - concord-client: renamed to `concord-tasks`;
 - concord-rpc: the KV store, heartbeat and secret gRPC services
 replaced with the REST API based services.
- 
+
+
 
 ## [0.69.0] - 2018-05-09